requirements.txt
flask
flask-cors
openai
python-dotenv
requests
psycopg2-binary
pyyaml
jinja2
bcrypt
PyJWT
